fix: Suppress embeds correctly
